+++ Woopsi/woopsi/graphicsport.cpp
@@ -161,7 +161,7 @@ void GraphicsPort::drawText(s16 x, s16 y, FontBase* font, u16 length, const char
 
 	if (_clipRect == NULL) {
 		// Draw all visible rects
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipText(x, y, font, length, string, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -344,7 +344,7 @@ void GraphicsPort::drawHorizLine(s16 x, s16 y, s16 width, u16 colour) {
 	if (_clipRect == NULL) {
 
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipHorizLine(x, y, width, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -364,7 +364,7 @@ void GraphicsPort::drawVertLine(s16 x, s16 y, s16 height, u16 colour) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipVertLine(x, y, height, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -384,7 +384,7 @@ void GraphicsPort::drawFilledRect(s16 x, s16 y, u16 width, u16 height, u16 colou
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipFilledRect(x, y, width, height, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -403,7 +403,7 @@ void GraphicsPort::drawFilledCircle(s16 x0, s16 y0, u16 radius, u16 colour) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipFilledCircle(x0, y0, radius, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -451,7 +451,7 @@ void GraphicsPort::drawRect(s16 x, s16 y, u16 width, u16 height, u16 colour) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipHorizLine(x, y, width, colour, _gadget->getVisibleRectCache()->at(i));						// Top
 			clipHorizLine(x, y + height - 1, width, colour, _gadget->getVisibleRectCache()->at(i));			// Bottom
 			clipVertLine(x, y + 1, height - 2, colour, _gadget->getVisibleRectCache()->at(i));				// Left
@@ -509,7 +509,7 @@ void GraphicsPort::drawBevelledRect(s16 x, s16 y, u16 width, u16 height, u16 shi
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipHorizLine(x, y, width, shineColour, _gadget->getVisibleRectCache()->at(i));							// Top
 			clipHorizLine(x, y + height - 1, width, shadowColour, _gadget->getVisibleRectCache()->at(i));			// Bottom
 			clipVertLine(x, y + 1, height - 2, shineColour, _gadget->getVisibleRectCache()->at(i));					// Left
@@ -535,16 +535,16 @@ void GraphicsPort::drawFilledXORRect(s16 x, s16 y, u16 width, u16 height) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 
 			// Draw all rows of rectangle
-			for (u8 j = 0; j < height; j++) {
+			for (s32 j = 0; j < height; j++) {
 				clipXORHorizLine(x, y + j, width, _gadget->getVisibleRectCache()->at(i));
 			}
 		}
 	} else {
 		// Draw single rectangle row by row
-		for (u8 j = 0; j < height; j++) {
+		for (s32 j = 0; j < height; j++) {
 			clipXORHorizLine(x, y + j, width, *_clipRect);	
 		}
 	}
@@ -561,7 +561,7 @@ void GraphicsPort::drawXORRect(s16 x, s16 y, u16 width, u16 height) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipXORHorizLine(x, y, width, _gadget->getVisibleRectCache()->at(i));						// Top
 			clipXORHorizLine(x, y + height - 1, width, _gadget->getVisibleRectCache()->at(i));			// Bottom
 			clipXORVertLine(x, y + 1, height - 2, _gadget->getVisibleRectCache()->at(i));				// Left
@@ -596,7 +596,7 @@ void GraphicsPort::drawBitmap(s16 x, s16 y, u16 width, u16 height, const u16* bi
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipBitmap(x, y, width, height, bitmap, bitmapX, bitmapY, bitmapWidth, bitmapHeight, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -625,7 +625,7 @@ void GraphicsPort::drawBitmap(s16 x, s16 y, u16 width, u16 height, const u16* bi
 
 	if (_clipRect == NULL) {
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipBitmap(x, y, width, height, bitmap, bitmapX, bitmapY, bitmapWidth, bitmapHeight, transparentColour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -646,7 +646,7 @@ void GraphicsPort::drawXORHorizLine(s16 x, s16 y, s16 width) {
 	if (_clipRect == NULL) {
 
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipXORHorizLine(x, y, width, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -668,7 +668,7 @@ void GraphicsPort::drawXORVertLine(s16 x, s16 y, s16 height) {
 	if (_clipRect == NULL) {
 
 		// Draw all visible rectangles
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipXORVertLine(x, y, height, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -862,7 +862,7 @@ void GraphicsPort::drawPixel(s16 x, s16 y, u16 colour) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rects
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipPixel(x, y, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
@@ -900,7 +900,7 @@ void GraphicsPort::drawLine(s16 x1, s16 y1, s16 x2, s16 y2, u16 colour) {
 
 	if (_clipRect == NULL) {
 		// Draw all visible rects
-		for (u8 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
+		for (s32 i = 0; i < _gadget->getVisibleRectCache()->size(); i++) {
 			clipLine(x1, y1, x2, y2, colour, _gadget->getVisibleRectCache()->at(i));
 		}
 	} else {
