+++ Woopsi/woopsi/textwriter.cpp
@@ -35,7 +35,7 @@ void TextWriter::drawString(FontBase* font, const char* string, u8 length, s16 x
 
 // Output a whole string to the display
 void TextWriter::drawString(u16* bitmap, u16 bitmapWidth, u16 bitmapHeight, FontBase* font, const char* string, u8 length, s16 x, s16 y, u16 clipX1, u16 clipY1, u16 clipX2, u16 clipY2) {
-	for (u8 i = 0; i < length; i++) {
+	for (u32 i = 0; i < length; i++) {
 		x = font->drawChar(bitmap, bitmapWidth, bitmapHeight, string[i], x, y, clipX1, clipY1, clipX2, clipY2);
 	}
 }
